+++ GUI/AchievementButton/AchievementsObjectivesMixin.lua
@@ -1,11 +1,19 @@
 -- [[ Namespaces ]] --
 local _, addon = ...;
+addon.GUI.AchievementsObjectives = {};
+local achievementsObjectives = addon.GUI.AchievementsObjectives;
+
+function achievementsObjectives.Load()
+	local frame = CreateFrame("Frame", "KrowiAF_AchievementsObjectives", nil, "KrowiAF_AchievementsObjectives_Template");
+	addon.GUI.AchievementsObjectives = frame;
+end
 
 AchievementsObjectivesMixin = {};
 
 AchievementsObjectivesMixin.Modes = {
     Criteria = 1,
-    Progressive = 2
+    Progressive = 2,
+	NoCriteria = 3
 }
 
 local criteriaTable, progressBarTable, miniTable, metaCriteriaTable = {}, {}, {}, {}
@@ -37,14 +45,14 @@ do -- Scripts
 	end
 
 	function KrowiAF_AchievementsObjectives_OnLoad(self)
-		local criteria = self:GetCriteria(1);
-		criteria.Label:SetText("- ");
-		self.TextCheckWidth = criteria.Label:GetStringWidth();
+		-- local criteria = self:GetTextCriteria(1);
+		-- criteria.Label:SetText("- ");
+		-- self.TextCheckWidth = criteria.Label:GetStringWidth();
 	end
 end
 
 do -- Reset objective types
-	function AchievementsObjectivesMixin:ResetCriteria()
+	function AchievementsObjectivesMixin:ResetTextCriteria()
 		AchievementButton_ResetTable(criteriaTable);
 	end
 
@@ -64,7 +72,7 @@ do -- Reset objective types
 	end
 
 	function AchievementsObjectivesMixin:ResetAll()
-		self:ResetCriteria();
+		self:ResetTextCriteria();
 		self:ResetProgressBars();
 		self:ResetMiniAchievements();
 		self:ResetMetas();
@@ -72,11 +80,11 @@ do -- Reset objective types
 end
 
 do -- Get objective types
-	function AchievementsObjectivesMixin:GetCriteria(index)
+	function AchievementsObjectivesMixin:GetTextCriteria(index)
 		if criteriaTable[index] then
 			return criteriaTable[index];
 		end
-		local frame = CreateFrame("FRAME", "KrowiAF_AchievementsObjectivesCriteria" .. index, self, "KrowiAF_AchievementCriteria_Template");
+		local frame = CreateFrame("FRAME", self:GetName() .. "TextCriteria" .. index, self, "KrowiAF_TextCriteria_Template");
 		AchievementFrame_LocalizeCriteria(frame);
 		criteriaTable[index] = frame;
 		return frame;
@@ -86,7 +94,7 @@ do -- Get objective types
 		if progressBarTable[index] then
 			return progressBarTable[index];
 		end
-		local frame = CreateFrame("STATUSBAR", "KrowiAF_AchievementsObjectivesProgressBar" .. index, self, "AchievementProgressBarTemplate");
+		local frame = CreateFrame("STATUSBAR", self:GetName() .. "ProgressBar" .. index, self, "AchievementProgressBarTemplate");
 		AchievementButton_LocalizeProgressBar(frame);
 		progressBarTable[index] = frame;
 		return frame;
@@ -96,7 +104,7 @@ do -- Get objective types
 		if miniTable[index] then
 			return miniTable[index];
 		end
-		local frame = CreateFrame("BUTTON", "KrowiAF_AchievementsObjectivesMiniAchievement" .. index, self, "KrowiAF_MiniAchievement_Template");
+		local frame = CreateFrame("BUTTON", self:GetName() .. "MiniAchievement" .. index, self, "KrowiAF_MiniAchievement_Template");
 		AchievementButton_LocalizeMiniAchievement(frame);
 		miniTable[index] = frame;
 		return frame;
@@ -106,7 +114,7 @@ do -- Get objective types
 		if metaCriteriaTable[index] then
 			return metaCriteriaTable[index];
 		end
-		local frame = CreateFrame("BUTTON", "KrowiAF_AchievementsObjectivesMeta" .. index, self, "KrowiAF_MetaCriteria_Template");
+		local frame = CreateFrame("BUTTON", self:GetName() .. "Meta" .. index, self, "KrowiAF_MetaCriteria_Template");
 		AchievementButton_LocalizeMetaAchievement(frame);
 		metaCriteriaTable[index] = frame;
 		return frame;
@@ -119,12 +127,16 @@ do -- Add objective types
 		metaCriteria:ClearAllPoints();
 		if index == 1 then
 			-- Anchor once all criteria are processed
+			metaCriteria:SetPoint("TOPLEFT", self, "TOPLEFT", 0, 0);
+			metaCriteria:SetPoint("RIGHT", self, "LEFT", self:GetWidth() / 2, 0);
 		elseif math.fmod(index, 2) == 0 then
 			local anchorMeta = self:GetMeta(index - 1);
-			metaCriteria:SetPoint("LEFT", anchorMeta, "RIGHT", 35, 0);
+			metaCriteria:SetPoint("LEFT", anchorMeta, "RIGHT", 0, 0);
+			metaCriteria:SetPoint("RIGHT", self, "RIGHT", 0, 0);
 		else
 			local anchorMeta = self:GetMeta(index - 2);
-			metaCriteria:SetPoint("TOPLEFT", anchorMeta, "BOTTOMLEFT", -0, 2);
+			metaCriteria:SetPoint("TOPLEFT", anchorMeta, "BOTTOMLEFT", 0, 2);
+			metaCriteria:SetPoint("RIGHT", self, "LEFT", self:GetWidth() / 2, 0);
 		end
 		local id, name, _, _, _, _, _, _, _, icon = GetAchievementInfo(assetId);
 		metaCriteria:Show();
@@ -154,88 +166,81 @@ do -- Add objective types
 		end
 	end
 
+	local progressBarOffset = 10;
 	function AchievementsObjectivesMixin:AddProgressBar(index, quantity, reqQuantity, quantityString)
 		local progressBar = self:GetProgressBar(index);
+		local extraHeight;
 		if index == 1 then
-			progressBar:SetPoint("TOP", self, "TOP", 4, -4);
+			progressBar:SetPoint("TOP", self, "TOP", 0, 0);
+			extraHeight = 5;
 		else
-			progressBar:SetPoint("TOP", self:GetProgressBar(index - 1), "BOTTOM", 0, 0);
+			progressBar:SetPoint("TOP", self:GetProgressBar(index - 1), "BOTTOM", 0, -progressBarOffset);
+			extraHeight = progressBarOffset;
 		end
 		progressBar.text:SetText(string.format("%s", quantityString));
 		progressBar:SetMinMaxValues(0, reqQuantity);
 		progressBar:SetValue(quantity);
 		progressBar:SetParent(self);
+		progressBar:SetHeight(16); -- Template is 14 but borders 16
 		progressBar:Show();
+
+		return progressBar:GetWidth(), progressBar:GetHeight() + extraHeight;
 	end
 
-	function AchievementsObjectivesMixin:AddCriteria(index, numCriteria, criteriaString, completed)
-		local criteria = self:GetCriteria(index);
+	function AchievementsObjectivesMixin:AddTextCriteria(index, numCriteria, criteriaString, completed)
+		local criteria = self:GetTextCriteria(index);
 		criteria:ClearAllPoints();
 		if index == 1 then
 			if numCriteria == 1 then
-				criteria:SetPoint("TOP", self, "TOP", -14);
+				criteria:SetPoint("TOP", self, "TOP", 0, 0);
 			else
-				criteria:SetPoint("TOPLEFT", self, "TOPLEFT", 0);
+				criteria:SetPoint("TOPLEFT", self, "TOPLEFT", 0, 0);
 			end
 		else
-			criteria:SetPoint("TOPLEFT", self:GetCriteria(index - 1), "BOTTOMLEFT", 0, 0);
+			criteria:SetPoint("TOPLEFT", self:GetTextCriteria(index - 1), "BOTTOMLEFT", 0, 0);
 		end
 		if self.Completed and completed then
-			-- criteria.Dash:SetTextColor(0, 0, 0, 1);
-			-- criteria.Dash:SetShadowOffset(0, 0);
+			criteria.Dash:SetTextColor(0, 0, 0, 1);
+			criteria.Dash:SetShadowOffset(0, 0);
 			criteria.Label:SetTextColor(0, 0, 0, 1);
 			criteria.Label:SetShadowOffset(0, 0);
 		elseif completed then
-			-- criteria.Dash:SetTextColor(0, 1, 0, 1);
-			-- criteria.Dash:SetShadowOffset(1, -1);
+			criteria.Dash:SetTextColor(0, 1, 0, 1);
+			criteria.Dash:SetShadowOffset(1, -1);
 			criteria.Label:SetTextColor(0, 1, 0, 1);
 			criteria.Label:SetShadowOffset(1, -1);
 		else
-			-- criteria.Dash:SetTextColor(0.6, 0.6, 0.6, 1);
-			-- criteria.Dash:SetShadowOffset(1, -1);
+			criteria.Dash:SetTextColor(0.6, 0.6, 0.6, 1);
+			criteria.Dash:SetShadowOffset(1, -1);
 			criteria.Label:SetTextColor(0.6, 0.6, 0.6, 1);
 			criteria.Label:SetShadowOffset(1, -1);
 		end
-		local stringWidth = 0;
-		local maxCriteriaContentWidth;
-		-- maxCriteriaContentWidth = ACHIEVEMENTUI_MAXCONTENTWIDTH - ACHIEVEMENTUI_CRITERIACHECKWIDTH;
+		local labelWidth = 0;
+		local maxLabelWidth = self:GetWidth() - criteria.Check:GetWidth();
 		if completed then
-			maxCriteriaContentWidth = ACHIEVEMENTUI_MAXCONTENTWIDTH - ACHIEVEMENTUI_CRITERIACHECKWIDTH;
-			criteria.Check:SetPoint("LEFT", 18, -3);
-			criteria.Label:SetPoint("LEFT", criteria.Check, "RIGHT", 0, 2);
-			-- criteria.Label:SetPoint("TOPLEFT", criteria.Check, "TOPRIGHT", 0, 0);
 			criteria.Check:Show();
-			-- criteria.Dash:Hide();
-			criteria.Label:SetText(criteriaString);
-			stringWidth = min(criteria.Label:GetStringWidth(), maxCriteriaContentWidth);
+			criteria.Dash:Hide();
 		else
-			maxCriteriaContentWidth = ACHIEVEMENTUI_MAXCONTENTWIDTH - self.TextCheckWidth;
-			criteria.Check:SetPoint("LEFT", 0, -3);
-			criteria.Label:SetPoint("LEFT", criteria.Check, "RIGHT", 5, 2);
-			-- criteria.Label:SetPoint("TOPLEFT", criteria.Check, "TOPRIGHT", -7, 0);
 			criteria.Check:Hide();
-			-- criteria.Dash:Show();
-			criteria.Label:SetText("- "..criteriaString);
-			-- criteria.Label:SetText(criteriaString);
-			stringWidth = min(criteria.Label:GetStringWidth() - self.TextCheckWidth, maxCriteriaContentWidth);	-- Don't want the "- " to be included in the width
+			criteria.Dash:Show();
 		end
-		-- criteria.Label:SetText(criteriaString);
-		-- stringWidth = min(criteria.Label:GetStringWidth(), maxCriteriaContentWidth);
-		local offset = 0;
-		if criteria.Label:GetWidth() > maxCriteriaContentWidth then
-			criteria.Label:SetWidth(maxCriteriaContentWidth);
-			offset = criteria.Label:GetHeight() / self.FontHeight + criteria.Label:GetHeight() / self.FontHeight * 5 - 5;
+		criteria.Label:SetText(criteriaString);
+		labelWidth = min(criteria.Label:GetStringWidth(), maxLabelWidth);
+		criteria.Label:SetWidth(labelWidth);
+		local height = 15;
+		if criteria.Label:GetStringWidth() > maxLabelWidth then
+			height = criteria.Label:GetHeight() + 5;
 		end
-		criteria:SetHeight(15 + offset);
 		criteria:SetParent(self);
+		criteria:SetWidth(labelWidth + criteria.Check:GetWidth());
+		criteria:SetHeight(height);
 		criteria:Show();
-		criteria:SetWidth(stringWidth + ACHIEVEMENTUI_CRITERIACHECKWIDTH);
 
-		return stringWidth;
+		return criteria:GetWidth(), criteria:GetHeight();
 	end
 end
 
-local achievements, rowOffset, columnOffset = {}, 8, 0;
+local achievements, rowOffset, columnOffset = {}, 8, 4;
 function AchievementsObjectivesMixin:DisplayProgressiveAchievement(id)
 	for i in next, achievements do
 		achievements[i] = nil;
@@ -248,19 +253,25 @@ function AchievementsObjectivesMixin:DisplayProgressiveAchievement(id)
 		prevId = GetPreviousAchievement(prevId);
 	end
 
-	local i = 0;
-	for index, achId in ipairs(achievements) do
+	local miniAchievementWidth = self:GetMiniAchievement(1):GetWidth();
+	local objectivesWidth = self:GetWidth();
+	local numColumns = floor(objectivesWidth / (miniAchievementWidth + columnOffset));
+	local numRows = ceil(#achievements / numColumns);
+	local width = min(numColumns, #achievements) * (miniAchievementWidth + columnOffset) - columnOffset;
+	local xOffset1 = (objectivesWidth - width) / 2;
+
+	for i, achId in ipairs(achievements) do
 		local _, _, points, _, _, _, _, _, _, icon = addon.GetAchievementInfo(achId);
-		local miniAchievement = self:GetMiniAchievement(index);
+		local miniAchievement = self:GetMiniAchievement(i);
 		miniAchievement:Show();
 		miniAchievement.Id = achId;
 		miniAchievement.Icon:SetTexture(icon);
-		if index == 1 then
-			miniAchievement:SetPoint("TOPLEFT", self, "TOPLEFT", -4, -4);
-		elseif mod(index, 6) == 1 then
-			miniAchievement:SetPoint("TOPLEFT", miniTable[index - 6], "BOTTOMLEFT", 0, -8);
+		if i == 1 then
+			miniAchievement:SetPoint("TOPLEFT", self, "TOPLEFT", xOffset1, 0);
+		elseif mod(i, numColumns) == 1 then
+			miniAchievement:SetPoint("TOPLEFT", self:GetMiniAchievement(i - numColumns), "BOTTOMLEFT", 0, -rowOffset);
 		else
-			miniAchievement:SetPoint("TOPLEFT", miniTable[index - 1], "TOPRIGHT", 4, 0);
+			miniAchievement:SetPoint("TOPLEFT", self:GetMiniAchievement(i - 1), "TOPRIGHT", columnOffset, 0);
 		end
 		if points > 0 then
 			miniAchievement.Points:SetText(points);
@@ -270,134 +281,67 @@ function AchievementsObjectivesMixin:DisplayProgressiveAchievement(id)
 			miniAchievement.Points:Hide();
 			miniAchievement.Shield:SetTexture("Interface/AchievementFrame/UI-Achievement-Progressive-Shield-NoPoints");
 		end
-		i = index;
 	end
-	self:SetHeight(math.ceil(i / 6) * (miniTable[i]:GetHeight() + rowOffset));
-	self:SetWidth(min(i, 6) * (miniTable[i]:GetWidth() + columnOffset));
+
+	self:SetHeight(numRows * (self:GetMiniAchievement(1):GetHeight() + rowOffset));
 	self.Mode = self.Modes.Progressive;
 end
 
-local FORCE_COLUMNS_MAX_WIDTH = 220;
-local FORCE_COLUMNS_MIN_CRITERIA = 20;
-local FORCE_COLUMNS_LEFT_OFFSET = -10;
-local FORCE_COLUMNS_RIGHT_OFFSET = 24;
-local FORCE_COLUMNS_RIGHT_COLUMN_SPACE = 150;
-
-function AchievementsObjectivesMixin:SetProgressBarAndTextPoints(progressBars, textStrings)
+function AchievementsObjectivesMixin:SetProgressBarAndTextPoints(numProgressBars, numTextCriteria)
 	-- If we have text criteria and progressBar criteria, display the progressBar criteria first and position the textStrings under them.
-	local criteria = self:GetCriteria(1);
-	criteria:ClearAllPoints();
-	if textStrings == 1 then
-		criteria:SetPoint("TOP", self:GetProgressBar(progressBars), "BOTTOM", -14, -4);
-	else
-		criteria:SetPoint("TOP", self:GetProgressBar(progressBars), "BOTTOM", 0, -4);
-		criteria:SetPoint("LEFT", self, "LEFT", 0, 0);
+	local criteria;
+	for i = 1, numTextCriteria do
+		criteria = self:GetTextCriteria(i);
+		criteria:ClearAllPoints();
+		if i == 1 then
+			criteria:SetPoint("TOP", self:GetProgressBar(numProgressBars), "BOTTOM", 0, -10);
+		else
+			criteria:SetPoint("TOP", self:GetTextCriteria(i - 1), "BOTTOM", 0, 0);
+		end
 	end
 end
 
-function AchievementsObjectivesMixin:ForceTwoRowsAdvanced(yOffset)
-	-- print(yOffset)
-	local numColumns = 2;
-	local xOffset = 0;
-	local maxColumnWidth = ACHIEVEMENTUI_MAXCONTENTWIDTH / numColumns;
-	local position = 0;
-	for i = 1, #criteriaTable do
-		position = position + 1;
-		if criteriaTable[i]:IsShown() then
-			if position == 2 and criteriaTable[i].Label:GetStringWidth() > maxColumnWidth then
-				position = position + 1;
-			end
-			if position > numColumns then
-				position = position - numColumns;
-				yOffset = yOffset - ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT;
-			end
-			criteriaTable[i]:ClearAllPoints();
-			if position == 1 then
-				xOffset = FORCE_COLUMNS_LEFT_OFFSET;
-			elseif position == 2 then
-				xOffset = FORCE_COLUMNS_RIGHT_OFFSET;
-			end
-			criteriaTable[i]:SetPoint("TOPLEFT", self, "TOPLEFT", (position - 1) * maxColumnWidth + xOffset, yOffset);
-			if criteriaTable[i].Label:GetStringWidth() > maxColumnWidth then
-				position = position + 1;
+function AchievementsObjectivesMixin:SetTextPoints(numTextCriteria, maxCriteriaWidth)
+	local columns = floor(self:GetWidth() / maxCriteriaWidth);
+
+	local truncate, flex;
+	if addon.Options.db.Achievements.Objectives.Force2Columns then
+		if columns < 2 and numTextCriteria >= addon.Options.db.Achievements.Objectives.Force2ColumnsThreshold then
+			columns = 2;
+			-- addon.Options.db.Achievements.Objectives.CriteriaBehaviour == 1 needs no additional code to overflow
+			if addon.Options.db.Achievements.Objectives.CriteriaBehaviour == 2 then -- Truncate
+				truncate = self:GetWidth() / 2;
+			elseif addon.Options.db.Achievements.Objectives.CriteriaBehaviour == 3 then
+				flex = true;
 			end
 		end
 	end
-	-- print(yOffset, ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT, abs(yOffset / ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT))
-	return abs(yOffset / ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT) + 1;
-end
 
-function AchievementsObjectivesMixin:SetTextPoints(progressBars, textStrings, maxCriteriaWidth, numCriteriaRows, addExtraCriteriaRow)
-	-- testing
-	-- FIND A WAY TO FORCE 2 COLUMNS AND SHOW THE TOO LONG ITEMS ON A SINGLE LINE
-	-- maxCriteriaWidth = min(maxCriteriaWidth, FORCE_COLUMNS_MAX_WIDTH)
-	-- testing
-	local yOffset = 0;
-	local numExtraCriteriaRows = 0;
-	local function AddExtraCriteriaRow()
-		numExtraCriteriaRows = numExtraCriteriaRows + 1;
-		yOffset = -numExtraCriteriaRows * ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT;
-	end
-	-- Figure out if we can make multiple columns worth of criteria instead of one long one
-	local numColumns = floor(ACHIEVEMENTUI_MAXCONTENTWIDTH / maxCriteriaWidth);
-	-- But if we have a lot of criteria, force 2 columns
+	local columnWidth = self:GetWidth() / columns;
 
-	-- print(ACHIEVEMENTUI_MAXCONTENTWIDTH, maxCriteriaWidth, numColumns, textStrings, FORCE_COLUMNS_MIN_CRITERIA, maxCriteriaWidth, FORCE_COLUMNS_MAX_WIDTH)
-
-	local forceColumns = false;
-	if numColumns == 1 and textStrings >= FORCE_COLUMNS_MIN_CRITERIA then
-		if maxCriteriaWidth <= FORCE_COLUMNS_MAX_WIDTH then
-			numColumns = 2;
-			forceColumns = true;
-			-- if top right criteria would run into the achievement shield, move them all down 1 row
-			-- this assumes description is 1 or 2 lines, otherwise this wouldn't be a problem
-			local numLines = self:GetParent().numLines;
-			if progressBars == 0 then
-				local ACHIEVEMENTUI_MAX_LINES_COLLAPSED = 3;
-				if self:GetCriteria(2).Label:GetStringWidth() > FORCE_COLUMNS_RIGHT_COLUMN_SPACE then
-					for i = 1, ACHIEVEMENTUI_MAX_LINES_COLLAPSED - numLines do
-						AddExtraCriteriaRow();
-					end
-				elseif self:GetCriteria(4).Label:GetStringWidth() > FORCE_COLUMNS_RIGHT_COLUMN_SPACE then
-					for i = 2, ACHIEVEMENTUI_MAX_LINES_COLLAPSED - numLines do
-						AddExtraCriteriaRow();
-					end
-				end
-			end
-		-- else
-		-- 	print("we just need 2 rows...")
-		-- 	return self:ForceTwoRowsAdvanced(yOffset);
-		end
-	end
-	if numColumns <= 1 then
-		return numCriteriaRows, numExtraCriteriaRows;
-	end
 	local rows = 1;
 	local position = 0;
-	for i=1, #criteriaTable do
+	local textCriteria;
+	for i = 1, numTextCriteria do -- The 1st one is already at its correct position
+		textCriteria = self:GetTextCriteria(i);
 		position = position + 1;
-		if ( position > numColumns ) then
-			position = position - numColumns;
+		if flex and position == 2 and textCriteria:GetWidth() > columnWidth then
+			position = position + 1;
+		end
+		if position > columns then
+			position = position - columns;
 			rows = rows + 1;
 		end
-		if ( rows == 1 ) then
-			criteriaTable[i]:ClearAllPoints();
-			local xOffset = 0;
-			if ( forceColumns ) then
-				if ( position == 1 ) then
-					xOffset = FORCE_COLUMNS_LEFT_OFFSET;
-				elseif ( position == 2 ) then
-					xOffset = FORCE_COLUMNS_RIGHT_OFFSET;
-				end
-			end
-			criteriaTable[i]:SetPoint("TOPLEFT", self, "TOPLEFT", (position - 1)*(ACHIEVEMENTUI_MAXCONTENTWIDTH/numColumns) + xOffset, yOffset);
-		else
-			criteriaTable[i]:ClearAllPoints();
-			criteriaTable[i]:SetPoint("TOPLEFT", criteriaTable[position + ((rows - 2) * numColumns)], "BOTTOMLEFT", 0, 0);
+		if truncate then
+			textCriteria:SetWidth(truncate);
+		end
+		textCriteria:ClearAllPoints();
+		textCriteria:SetPoint("TOPLEFT", self, "TOPLEFT", (position - 1) * columnWidth, -(rows - 1) * textCriteria:GetHeight());
+		if flex and textCriteria:GetWidth() > columnWidth then
+			position = position + 1;
 		end
 	end
-	numCriteriaRows = ceil(numCriteriaRows / numColumns);
-	return numCriteriaRows, numExtraCriteriaRows;
+	return self:GetTextCriteria(1):GetTop() - self:GetTextCriteria(numTextCriteria):GetBottom();
 end
 
 function AchievementsObjectivesMixin:DisplayCriteria(id)
@@ -407,51 +351,54 @@ function AchievementsObjectivesMixin:DisplayCriteria(id)
 
 	local numCriteria = GetAchievementNumCriteria(id);
 	if numCriteria == 0 then
-		self.Mode = self.Modes.Criteria;
-		self:SetHeight(0);
+		self.Mode = self.Modes.NoCriteria;
 		return;
 	end
 
-	local stringWidth;
+	local progressBarWidth, progressBarHeight;
+	local totalProgressBarHeight = 0;
+	local textCriteriaWidth, textCriteriaHeight;
+	local totalTextCriteriaHeight = 0;
 	local numCriteriaRows = 0;
 	local numExtraCriteriaRows = 0;
 
-	local textStrings, progressBars, metas = 0, 0, 0;
+	local numTextCriteria, numProgressBars, numMetas = 0, 0, 0;
 	local maxCriteriaWidth = 0;
 	for i = 1, numCriteria do
 		local criteriaString, criteriaType, completed, quantity, reqQuantity, _, flags, assetID, quantityString = GetAchievementCriteriaInfo(id, i);
+		flags = addon.Objects.Flags:New(flags);
 		if criteriaType == CRITERIA_TYPE_ACHIEVEMENT and assetID then
-			metas = metas + 1;
-			self:AddMeta(metas, completed, assetID)
-		elseif ( bit.band(flags, EVALUATION_TREE_FLAG_PROGRESS_BAR) == EVALUATION_TREE_FLAG_PROGRESS_BAR ) then
-			progressBars = progressBars + 1;
-			self:AddProgressBar(progressBars, quantity, reqQuantity, quantityString)
+			numMetas = numMetas + 1;
+			self:AddMeta(numMetas, completed, assetID)
+		elseif flags.IsCriteriaProgressBar then
+			numProgressBars = numProgressBars + 1;
+			progressBarWidth, progressBarHeight = self:AddProgressBar(numProgressBars, quantity, reqQuantity, quantityString);
+			totalProgressBarHeight = totalProgressBarHeight + progressBarHeight;
 			numCriteriaRows = numCriteriaRows + 1;
 		else
-			textStrings = textStrings + 1;
-			stringWidth = self:AddCriteria(textStrings, numCriteria, criteriaString, completed);
-			maxCriteriaWidth = max(maxCriteriaWidth, stringWidth + ACHIEVEMENTUI_CRITERIACHECKWIDTH);
+			numTextCriteria = numTextCriteria + 1;
+			textCriteriaWidth, textCriteriaHeight = self:AddTextCriteria(numTextCriteria, numCriteria, criteriaString, completed);
+			maxCriteriaWidth = max(maxCriteriaWidth, textCriteriaWidth);
+			totalTextCriteriaHeight = totalTextCriteriaHeight + textCriteriaHeight;
 			numCriteriaRows = numCriteriaRows + 1;
 		end
 	end
-	if textStrings > 0 and progressBars > 0 then
-		self:SetProgressBarAndTextPoints(progressBars, textStrings);
-	elseif ( textStrings > 1 ) then
-		numCriteriaRows, numExtraCriteriaRows = self:SetTextPoints(progressBars, textStrings, maxCriteriaWidth, numCriteriaRows)
-	end
-	numCriteriaRows = numCriteriaRows + numExtraCriteriaRows;
-	local firstMetaCriteria = self:GetMeta(1);
-	if firstMetaCriteria.Id then
-		local yOffsetMeta = -8 - numCriteriaRows * ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT;
-		if metas == 1 then
-			firstMetaCriteria:SetPoint("TOP", self, "TOP", 0, yOffsetMeta);
-		else
-			firstMetaCriteria:SetPoint("TOPLEFT", self, "TOPLEFT", 20, yOffsetMeta);
+	local height = 1;
+	if numProgressBars > 0 then
+		self:SetProgressBarAndTextPoints(numProgressBars, numTextCriteria);
+		height = totalProgressBarHeight + totalTextCriteriaHeight;
+	elseif numTextCriteria > 0 then
+		height = totalTextCriteriaHeight;
+		height = self:SetTextPoints(numTextCriteria, maxCriteriaWidth);
+		if numMetas > 0 then
+			self:GetMeta(1):ClearAllPoints();
+			self:GetMeta(1):SetPoint("TOP", self:GetTextCriteria(numTextCriteria), "BOTTOM", 0, 0);
+			self:GetMeta(1):SetPoint("LEFT", self, "LEFT", 0, 0);
+			self:GetMeta(1):SetPoint("RIGHT", self, "LEFT", self:GetWidth() / 2, 0);
+			height = height + ceil(numMetas / 2) * ACHIEVEMENTBUTTON_METAROWHEIGHT;
 		end
-	end
-	local height = ceil(metas / 2) * ACHIEVEMENTBUTTON_METAROWHEIGHT + numCriteriaRows * ACHIEVEMENTBUTTON_CRITERIAROWHEIGHT;
-	if ( metas > 0 or progressBars > 0 ) then
-		height = height + 10;
+	else
+		height = ceil(numMetas / 2) * ACHIEVEMENTBUTTON_METAROWHEIGHT;
 	end
 	self:SetHeight(height);
 	self.Mode = self.Modes.Criteria;