+++ GUI/AchievementsFrame.lua
@@ -34,6 +34,8 @@ function achievementsFrame:Load()
 		frame:ForceUpdate();
 	end); -- Issue #3: Fix
 
+	frame.ScrollBarStep = scrollBar:GetValueStep();
+
 	addon.GUI.AchievementsFrame = frame; -- Overwrite with the actual frame since all functions are injected to it
 end
 
@@ -102,7 +104,8 @@ function KrowiAF_AchievementsFrameMixin:Update()
 
 	-- Let's try just always hide it. When switching tabs and the new tab has no achievement selected, this line or ClearSelection is not called
 	-- if selectedAchievement then
-		self.AchievementsObjectives:Hide();
+	addon.GUI.AchievementsObjectives:Hide();
+	scrollFrame.ScrollBar:SetValueStep(self.ScrollBarStep);
 	-- end
 
 	local displayedHeight = 0;
@@ -155,7 +158,7 @@ function KrowiAF_AchievementsFrameMixin:ExpandSelection(button)
 end
 
 function KrowiAF_AchievementsFrameMixin:ClearSelection()
-	self.AchievementsObjectives:Hide();
+	addon.GUI.AchievementsObjectives:Hide();
 	local buttons = self.ScrollFrame.buttons;
 	for _, button in next, buttons do
 		button.selected = nil;
@@ -204,7 +207,8 @@ end
 	end
 end
 
- -- When the selection is already visible, adjust it so it fits
+-- When the selection is already visible, adjust it so it fits
+
 function KrowiAF_AchievementsFrameMixin:AdjustSelection()
 	local scrollFrame = self.ScrollFrame;
 	local scrollBar = scrollFrame.ScrollBar;
@@ -237,6 +241,7 @@ function KrowiAF_AchievementsFrameMixin:AdjustSelection()
 	if newHeight then
 		local _, maxVal = scrollBar:GetMinMaxValues();
 		newHeight = min(newHeight, maxVal);
+		scrollBar:SetValueStep(1);
 		scrollBar:SetValue(newHeight);
 	end
 end
@@ -270,8 +275,8 @@ function KrowiAF_AchievementsFrameMixin:ForceUpdate(toTop) -- Issue #3: Fix
 	selectedTab.SelectedAchievement = addon.Filters.GetHighestAchievementWhenCollapseSeries(selectedTab.Filters, selectedTab.SelectedAchievement);
 
 	-- Issue #8: Broken
-	self.AchievementsObjectives:Hide();
-	self.AchievementsObjectives.id = nil;
+	addon.GUI.AchievementsObjectives:Hide();
+	addon.GUI.AchievementsObjectives.id = nil;
 
 	local buttons = self.ScrollFrame.buttons;
 	for _, button in next, buttons do