+++ Options/Defaults.lua
@@ -60,7 +60,12 @@ options.Defaults = {
         },
         Achievements = {
             Compact = false,
-            ButtonCollapsedHeight = ACHIEVEMENTBUTTON_COLLAPSEDHEIGHT or 84 -- can return nil
+            ButtonCollapsedHeight = ACHIEVEMENTBUTTON_COLLAPSEDHEIGHT or 84, -- can return nil
+            Objectives = {
+                Force2Columns = true,
+                Force2ColumnsThreshold = 20,
+                CriteriaBehaviour = 2
+            }
         },
         RightClickMenu = {
             ShowButtonOnAchievement = false,